Changes per Code Review
