+++ src/renderer/hooks/papi-hooks/index.ts
@@ -4,12 +4,14 @@ import useEventAsync from '@renderer/hooks/papi-hooks/use-event-async.hook';
 import useDataProvider from '@renderer/hooks/papi-hooks/use-data-provider.hook';
 import useData from '@renderer/hooks/papi-hooks/use-data.hook';
 import useSetting from '@renderer/hooks/papi-hooks/use-setting.hook';
+import useProjectDataProvider from './use-project-data-provider.hook';
 
 // Declare an interface for the object we're exporting so that JSDoc comments propagate
 export interface PapiHooks {
   usePromise: typeof usePromise;
   useEvent: typeof useEvent;
   useEventAsync: typeof useEventAsync;
+  useProjectDataProvider: typeof useProjectDataProvider;
   useDataProvider: typeof useDataProvider;
   /** JSDOC DESTINATION UseDataHook */
   useData: typeof useData;
@@ -23,6 +25,7 @@ const papiHooks: PapiHooks = {
   usePromise,
   useEvent,
   useEventAsync,
+  useProjectDataProvider,
   useDataProvider,
   useData,
   useSetting,
